Server/requirements.txt
Django
djangorestframework
psycopg2-binary
pdoc3
djangochannelsrestframework
uvicorn
